generall/improoved code, implement code review rejects
