+++ src/Assets/Styles/Components/Appheader/_AppHeader.scss
@@ -2,9 +2,6 @@
     height: $app-header-height;
     padding: 0 20px;
     z-index: 10;
-    // position: fixed;
-    // top: 0px;
-    // width: 100%;
     background-color: white;
     .right-nav {
         right: $xLarge-space;
