+++ src/Components/Dynamic/CustomTablPanel.tsx
@@ -1,0 +1,27 @@
+import Box from "@mui/material/Box";
+
+interface TabPanelProps {
+    children?: React.ReactNode;
+    index: number;
+    value: number;
+}
+
+export const CustomTabPanel = (props: TabPanelProps) => {
+    const { children, value, index, ...other } = props;
+
+    return (
+        <div
+            key={index}
+            role="tabpanel"
+            hidden={value !== index}
+            id={`simple-tabpanel-${index}`}
+            {...other}
+        >
+            {value === index && (
+                <Box sx={{ p: 3 }}>
+                    <div>{children}</div>
+                </Box>
+            )}
+        </div>
+    );
+};
