+++ src/Pages/Checkout.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Checkout = () => {
+    return <div>Checkout</div>;
+};
+
+export default Checkout;
